chore: dissolve `bech32`
